+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [UNRELEASED]
+
+### Changed
+
+- The request handler used to generate the response must implement `Interop\Http\Server\RequestHandlerInterface`. Removed support for callables.
+
+### Removed
+
+- Removed `arguments()` option.
+
 ## [0.6.0] - 2017-11-13
 
 ### Changed
@@ -52,6 +62,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 First version
 
+[UNRELEASED]: https://github.com/middlewares/shutdown/compare/v0.6.0...HEAD
 [0.6.0]: https://github.com/middlewares/shutdown/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/middlewares/shutdown/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/middlewares/shutdown/compare/v0.3.0...v0.4.0